Designed the pid controller in simulink
I've replaced the custom built C module "PidTests" with a simulink model
I've designed to use the 6 dof block to properly model the quadcopter.
The closed loop control structure that deals with processing the euler
angles and executing the appropriate pid algorithms have also been
implemented.

At this time, I've set up the model to output a crude simulation result
that tracks the velocity, angular velocity, displacement and angular
position of the quadcopter.

Note that the values used in the model describing quadcopter parameters
such as weight, thrust, torque... have not been adjusted to deal with
the new frame.