Created the driver for the new IMU chip.
Wrote and successfully tested an interrupt based SPI driver for the
bmx160. The driver initialises the chip and allows collection of the
accelerometer, gyroscope and magnetometer data.

On top of this, the processor clock has been sped up to 20MHz and the
spi clock runs at 10 MHz.

Although the driver has been tested, it is not yet integrated with the
rest of the code.