Made the pwmchip I2C driver interrupt based
There are now no blocking calls associated with the pwm chip. The I2C
driver runs at 100 KHz.

Although the driver has been tested and functions, it is not yet
integrated with the rest of the code. With this commit, the code will
still compile, but will not fly a quadcopter properly until changes are
made to the controller module.