Centralized the interrupt functionality relevant to the Pilot instructions module inside it.
Previous to this change, the external interrupts responsible for reading
the pwm signals coming from the receiver were initialized in main and
the ISRs existed in "interrupts.c". With this change, The
PilotInstructions module takes care of initializing its interrupts and
holds the ISRs, thus removing the need for an interrupt module and
cleaning up main.