Integrated multiple modules into Controller to add functionality up to gathering and fusing all data from the sensors.
This is a large and important commit. As of this commit, Controller's
state machine reflects the design from the flowchart and can effectively
collect Pilot instructions, altitude, and IMU data from the relevant
hardware. This is all done in a real time loop (20ms period at the
moment).

The other new change is the introduction of the MahonyAHRS module, which
is responsible for fusing together the measurements made by the IMU into
a quaternion representing the orientation of the quad in space. That
quaternion is subsequently converted into XYZ Euler angles and their
derivatives, as the PID algorithm, which will be added in a later
commit, has been designed to run on Euler angles.

Although I plan to put the calibration module back in, I removed it from
main for the moment, as it will need to be rethought and reintegrated.

In the process of integrating the modules together, I decided to move
the ImuData struct definition to the common module, so that the struct
could be passed around easier. The changes in the project file just
reflect the addition of new files.