Fixed important bug in Altitude.c
Previous to this change, the echo pulse of the ultrasonic distance
sensor would only be properly measured if it represented an altitude
below 50 cm and if the timer that measured it did not wrap around.

With this change, the timer that measures the echo pulse has been
prescaled by a factor of 64, so all altitudes in the sensor's range can
be measured. Also, the edge case of the timer overflowing during the
measurement has been covered and tested.