Rebuilt the calibration module
Now, the pilot has the ability to request a calibration of the
ultrasonic distance sensor and of his own controls. That calibration
gets stored in eeprom and is used to adjust the modules every time the
board is powered on.

The other thing done by calibration, on each power cycle, is ensure that
the pilot's command of 0 yaw corresponds to the direction the Quad is
pointing, and not necessarily north.

For all this to happen, LoadCalibration functions had to be added to the
Altitude and PilotInstructions modules. Also, main now has to call the
appropriate calibration function. The changes in the project file
reflect the addition of a new module.