Increased control loop frequency from 50 to 200 Hz
Doing some testing, it was observed that it takes roughly 3 ms for all
the math to be done in a single control loop iteration. Since the Imu
supports 200 Hz gyro and acc data rates, it is certainly possible to
output results at this speed.

Although it was not extensively tested, increasing the control loop
frequency can only be beneficial, as finer corrections can be made more
often, resulting in smoother and more stable characteristics.

The Imu driver had to be modified to increase the gyroscope and
accelerometer ODRs so as to support the increased data rate. The Mahony
filter was updated with the new control loop frequency, as was the macro
in Common. The Pwm chip driver had to be modified to reduce the period
of the pwm signals sent to the receiver. The Altitude module had to
modified aswell. Since the ultrasonic sensor that is the backbone of
the module cannot sample data at the new speed, the driver had to be
updated so that measurements are not interrupted while they are in
progress.