Reduced the value associated with MOTOR_VALUE_NO_SPIN
The old value was slightly too high, so the motors sometimes refused to start (there is a safety mechanism in the esc that does not allow the motors to start spinning unless a state of near 0 throttle is entered first). With the new value of 0.5, this problem has not been detected.