Completed the calibration sequence of the magnetometer and gyroscope
As of this commit, there is a sequence of commands (that will be explained in the readme) that the pilot can use on startup to calibrate the magnetometer and gyroscope. The magnetometer calibration has effectively been shown to eliminate the hard iron biases present on the quad, which allows the sensor fusion algorithms to properly get the information about earth's magnetic field. The gyro calibration has also been effectively shown to eliminate the systematic error associated with the sensor.

Aside from the above changes, the Calibration module now, once again, leverages the Leds to indicate to the pilot what stage of Calibration is currently in progress.

The common module now contains a couple more macros and includes that were needed by multiple other modules. The Controller module now initialises the leds module. To the Imu module, a function was added to load the calibration; that function gets called by the Calibration module. The changes to the Leds module just reflect the new PCB (it was not updated since hardware revision 2). Lastly, the change in the project file reflects that the optimization level was increased from optimize for debugging to optimization level 1. That was done as the delay function is used multiple times by the Calibration module, and that function requires compiler optimizations to work properly.