Updated the Pid module with the latest gains and covered yaw edge case
A few commits ago, the control loop was sped up to 200 Hz. As a result, the matlab model was updated and the optimal gains changed. Those latest gains are now reflected in the Pid module.

Furthermore, it was discovered that there are possible discontinuities in the way the Pid algorithms treated the z axis. For example, if the pilot commanded -170 degrees while the quad was at 175 degrees, the old algorithm would attempt to "go around the long way". Such cases are now accounted for. I'll not that such cases are not possible in the x or y axes because the quad should never be close to "flipping around" anyway.