Used non magnetometer API of Mahony rather than full version
I discovered that I live so far north that the earth's magnetic field around here is not actually perpendicular to gravity. This really confused the Mahony algorithm. In the future, I will make some software to correct the magnetometer readings for this, but for the time being, I'm moving forward without accounting for Earth's magnetic field.