Added initialisation delays
There was a bug plaging this project for many weeks. It was eventually
understood to  be caused by a failure to properly initialize the
accelerometer, which seems to be a result of not giving the SPI module
enough time to initialise before putting it in use. This problem was
solved by adding a delay to allow that time. A delay was also added to
the very first line of the main function, for good measure.