More little things in this commit
I've disabled the finding of the initial azimuth in the calibration since the
magnetometer is not currently being used in the sensor fusion algorithm.

I've also fixed the signs of the roll pitch and yaw adjustements in the pid
module so they would correspond with reality. I also decided to go back to
proportional control (rather than proportional squared) because results seem
fine with it and I don't want to complicate things.

I've fixed the calculation of the euler angle derivatives to correspond with
the new Z,Y,X sequence.

The change in the project file is there because I've removed the
double-promotion warning. That was done because it turns out that the avr-gcc
compiler demotes all doubles to floats anyway.