final commit before release
The big change in this commit is the addition of a lowPass filter to
smooth altitude data as it was found to contain outliers that threw off
the derivative wildly.

The other major change is a further optimization of the pid gains. The
new gains were found by the matlab model to give better results.

All other changes are typos that were fixed and comments that were
added.

The Readme was modified to reflect the current state of the project.