Quad now uses altimeter to read altitudes past what the ultrasonic can do.
As of this commit, the altitude module reads the altimeter chip constantly and decides when to use the altimeter over the ultrasonic distance sensor. This change is opaque to all other modules though, whom simply receive an altitude. The advantage here is that while the ultrasonic distance sensor can only take measurements up to 4 meters, the altimeter has pretty much no cap on max altitude, so pilots can fly much higher.

The changes in the I2C module are there to support communications with the altimeter (which uses a different bus than the PWM chip). As a result, the API of the I2C driver had to change and the calls in PwmChip.c had to change accordingly. The changes in Calibration reflect the fact that we now have a new sensor to calibrate. Finally the changes in Controller are present because lowPass filtering of the ultrasonic distance sensor was moved from Controller to the altitude module.